Added Multi Scale Fourier Feature Network as per Issue #498
Considering the amount of code rewritten, it makes more sense to define it as a struct or purely define one ff layer that encapsulates a fourier feature encoding and dense layers and then have a different file where users can create msff networks using FastChains or a new file for Networks that are defined using general FastChains. 

Addtionally, bias is not included because of my lack of understanding of including it in a solitary function. 
It has not been rigirously tested, although it seems to compile.  I did get an inexact error that I couldn't debug when trying to run some operations, I would need some help to tackle the issue now that I know the general way that the network can be implemented.